Add basic architecture for implementing instructions
Not 100% sure if this is the best way or not... but it's what I have for now. Hopefully getting experience implementing the instructions will lead to the optimal solution.

Another one that comes to mind is a React/Redux reducer. That would look like a function that receives "actions" (load, tick, etc), and returns a new object after each.